Update toast notifications for 64 bit
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/branches/gnustep_testplant_branch@40260 72102866-910b-0410-8b05-ffd578937521

+++ Source/win32/MSUserNotifications/ToastNotifications/ToastNotifications/ToastNotifications/ToastNotifications.cpp
@@ -249,7 +249,7 @@ HRESULT CToastNotificationsApp::CreateToastXml(_In_ IToastNotificationManagerSta
           L"      "
         };
 
-        UINT32 textLengths[] = { wcslen(notificationTitle), wcslen(notificationDescription), 6 };
+        UINT32 textLengths[] = { (UINT32)wcslen(notificationTitle), (UINT32)wcslen(notificationDescription), 6 };
         hr = SetTextValues(textValues, 3, textLengths, *inputXml);
       }
